+++ modules/control/proc/zoe/zoedepth/models/zoedepth/zoedepth_v1.py
@@ -26,12 +26,12 @@ import itertools
 
 import torch
 import torch.nn as nn
-from ..depth_model import DepthModel
-from ...base_models.midas import MidasCore
-from ...layers.attractor import AttractorLayer, AttractorLayerUnnormed
-from ...layers.dist_layers import ConditionalLogBinomial
-from ...layers.localbins_layers import Projector, SeedBinRegressor, SeedBinRegressorUnnormed
-from ...model_io import load_state_from_resource
+from modules.control.proc.zoe.zoedepth.models.depth_model import DepthModel
+from modules.control.proc.zoe.zoedepth.models.base_models.midas import MidasCore
+from modules.control.proc.zoe.zoedepth.models.layers.attractor import AttractorLayer, AttractorLayerUnnormed
+from modules.control.proc.zoe.zoedepth.models.layers.dist_layers import ConditionalLogBinomial
+from modules.control.proc.zoe.zoedepth.models.layers.localbins_layers import Projector, SeedBinRegressor, SeedBinRegressorUnnormed
+from modules.control.proc.zoe.zoedepth.models.model_io import load_state_from_resource
 
 
 class ZoeDepth(DepthModel):