+++ modules/control/processor.py
@@ -18,8 +18,6 @@ processors = [
     # pose
     'OpenPose',
     'DWPose',
-    'MediaPipe Face',
-    'DWPose (ONNX)',
     'RTMW',
     'RTMO',
     'ViTPose',
@@ -31,8 +29,7 @@ processors = [
     'HED',
     'PidiNet',
     'MLSD',
-    'TEED',
-    'Anyline',
+    'Anyline (Legacy)',
     # depth
     'Midas Depth Hybrid',
     'Leres Depth',
@@ -47,16 +44,20 @@ processors = [
     'Marigold Depth LCM',
     'Lotus Depth',
     # normal
-    'Normal Bae',
+    'Normal Bae (Legacy)',
     'DSINE',
     'StableNormal',
     'Marigold Normals',
     # segmentation
-    'SegmentAnything',
-    'SAM 2.1',
+    'SegmentAnything 1.0',
+    'SegmentAnything 2.1',
     'OneFormer',
     # other
     'Shuffle',
+    # legacy
+    'MediaPipe Face (Legacy)',
+    'DWPose (Legacy)',
+    'TEED (Legacy)',
 ]